// src/app/core/api/swapi/models/starships.swapi.ts
/**
 * Raw data contract received directly from SWAPI for starships (https://swapi.info/api/starships)
 */
export interface SwapiStarshipDto {
  name: string;
  model: string;
  manufacturer: string;
  cost_in_credits: string;
  length: string;
  max_atmosphering_speed: string;
  crew: string;
  passengers: string;
  cargo_capacity: string;
  consumables: string;
  hyperdrive_rating: string;
  MGLT: string;
  starship_class: string;
  pilots: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
}